TASK/GLUE/requirements.txt
visualdl
tensorboardX
paddlenlp
tqdm
scipy
sklearn
easydict
sentencepiece